backend/pyproject.toml
[tool.poetry]
name = "reflect-backend"
version = "0.1.0"
description = ""
authors = ["Priyam Srivastava <ipriyam26@gmail.com>"]
readme = "README.md"
packages = [{include = "reflect_backend"}]

[tool.poetry.dependencies]
python = "^3.11"
openai = "^0.28.0"
langchain = "^0.0.300"
tiktoken = "^0.5.1"
python-dotenv = "^1.0.0"
fastapi = "^0.103.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
jinja2 = "^3.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
